+++ raycaster/src/world.rs
@@ -1,8 +1,8 @@
 use crate::color::*;
 
 pub struct World {
-    width: Option<i32>,
-    height: Option<i32>,
+    width: Option<i64>,
+    height: Option<i64>,
     map: Option<Vec<u8>>,
 }
 
@@ -16,7 +16,7 @@ impl World {
         }
     }
 
-    pub fn init(&mut self, width: i32, height: i32, map: Vec<u8>){
+    pub fn init(&mut self, width: i64, height: i64, map: Vec<u8>){
         self.width = Some(width);
         self.height = Some(height);
         self.map = Some(map);
@@ -26,7 +26,7 @@ impl World {
         self.width.is_some() && self.height.is_some() && self.map.is_some()
     }
 
-    pub fn is_collision(&self, xpos: i32, ypos: i32) -> bool {
+    pub fn is_collision(&self, xpos: i64, ypos: i64) -> bool {
         // For now we will treat the uninitialized map as out of bounds
         if !self.is_initialized() {
             return true;
@@ -45,7 +45,7 @@ impl World {
     }
 
     // TODO: This needs to have bounds checking
-    pub fn get_cell(&self, xpos: i32, ypos: i32) -> u8 {
+    pub fn get_cell(&self, xpos: i64, ypos: i64) -> u8 {
         let width = self.width.unwrap();
         let index = ypos * width + xpos;
         self.map.as_ref().unwrap()[index as usize]
